chore: move serialize functionality into the document and index files
